feat: Add watchlist functionality and show snackbar message
This commit introduces the ability to add coins to the watchlist from the search results. It also implements a snackbar to display a confirmation message when a coin is added to the watchlist. The `SearchAdapter` was updated to include a listener for adding coins to the watchlist. A new `showMessage` method was added to the `HomeFragment` to show snackbars. Finally an action listener was added to the `extendedFab`.