feat: Configure release signing with keystore properties
This commit configures the application to use a `keystore.properties` file for release signing. It defines a `release` signing configuration, specifying the key alias, key password, store file, and store password. The debug and release build types are set to use this signing configuration. Minification and resource shrinking are enabled for the release build.